Default traits to empty array in useFlags
The typescript definitions for the `useFlags` mark the traits array as optional. However, if you omit the array when using the hook, you'll be faced with the following warning which comes from the `flagsAsArray` function.

> 'Flagsmith: please supply an array of strings or a single string of flag keys to useFlags'

Changing the traits to default to an empty array in the function signature would remove the error and still keep the types the same. We just have to change `?:` to `:` otherwise TS will complain that you can't have a `?` and a default value.